[Fix] Fix another stupid warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5876 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/ModInstrument.cpp
@@ -129,7 +129,7 @@ int32 InstrumentEnvelope::GetValueFromPosition(int position, int32 rangeOut, int
 void InstrumentEnvelope::Sanitize(uint8 maxValue)
 //-----------------------------------------------
 {
-	LimitMax(nNodes, MAX_ENVPOINTS);
+	LimitMax(nNodes, uint32(MAX_ENVPOINTS));
 	Ticks[0] = 0;
 	for(uint32 i = 1; i < nNodes; i++)
 	{
